requirements.txt
cymysql==0.9.4
Django==2.0
django-autocomplete-light==3.2.10
django-autoslug==1.9.3
django-cities-light==3.3.0
django-crispy-forms==1.7.0
django-cymysql==2.0.0
django-filter==1.1.0
django-widget-tweaks==1.4.1
gunicorn==19.7.1
html5lib==1.0b10
httplib2==0.10.3
Pillow==5.0.0
progressbar2==3.34.3
PyPDF2==1.26.0
python-utils==2.2.0
pytz==2017.3
reportlab==3.4.0
six==1.11.0
Unidecode==0.4.21
webencodings==0.5.1
xhtml2pdf==0.2b1
xlrd==1.1.0
